removed filterCart from hook to be standalone util function
