+++ use-shopping-cart/src/index.test.js
@@ -625,48 +625,3 @@ describe('loadCart()', () => {
     expect(cart.current.cartCount).toEqual(8)
   })
 })
-
-describe('filterCart', () => {
-  const wrapper = createWrapper()
-  const cart = renderHook(() => useShoppingCart(), { wrapper }).result
-  it('filter out all cart items that fail filter callback', async () => {
-    const cartDetails = {
-      'test-sku-123': {
-        sku: 'test-sku-123',
-        name: 'Bananas',
-        image: 'https://blah.com/banana.avif',
-        price: 200,
-        currency: 'USD',
-        value: 400,
-        quantity: 2,
-        formattedValue: '$4.00'
-      },
-      'test-sku-234': {
-        sku: 'test-sku-234',
-        name: 'Oranges',
-        image: 'https://blah.com/orange.avif',
-        currency: 'USD',
-        price: 250,
-        value: 1000,
-        quantity: 4,
-        formattedValue: '$10.00'
-      }
-    }
-
-    const mockFilter = (item) => item.value < 500
-    const filteredCart = await cart.current.filterCart(cartDetails, mockFilter)
-
-    expect(filteredCart).toEqual({
-      'test-sku-123': {
-        sku: 'test-sku-123',
-        name: 'Bananas',
-        image: 'https://blah.com/banana.avif',
-        price: 200,
-        currency: 'USD',
-        value: 400,
-        quantity: 2,
-        formattedValue: '$4.00'
-      }
-    })
-  })
-})
